Fix entities relations, weeklys jobs, add team to wishlists

+++ nest-api/src/wishlists/wishlists.controller.ts
@@ -1,7 +1,8 @@
-import { BadRequestException, Body, Controller, Get, Param, Post, Query, UseGuards } from '@nestjs/common';
+import { BadRequestException, Body, Controller, Get, NotFoundException, Param, Post, Query, UseGuards } from '@nestjs/common';
 import { ApiOperation, ApiParam, ApiQuery, ApiTags } from '@nestjs/swagger';
 import JwtAuthenticationGuard from 'src/auth/jwt-authentication.guard';
 import { RaidDifficulty } from 'src/raids/raid.entity';
+import { TeamsService } from 'src/teams/teams.service';
 import { WishDto } from './whish.dto';
 import { WishlistsService } from './wishlists.service';
 
@@ -9,7 +10,8 @@ import { WishlistsService } from './wishlists.service';
 @Controller('wishlists')
 export class WishlistsController {
     constructor(
-        private readonly wishlistsService: WishlistsService
+        private readonly wishlistsService: WishlistsService,
+        private readonly teamsService: TeamsService
     ) {}
 
     @UseGuards(JwtAuthenticationGuard)
@@ -18,9 +20,11 @@ export class WishlistsController {
     @ApiQuery({ name: 'difficulty', enum: RaidDifficulty, required: false })
     @ApiQuery({ name: 'encounter', required: false, type: "number" })
     @ApiQuery({ name: 'raid', required: false, type: "number" })
+    @ApiQuery({ name: 'team', type: 'number' })
     @ApiOperation({ summary: 'Retrieve wishlist of a given character' })
     @Get()
     async getAll(
+        @Query('team') teamId: number,
         @Query('character') character: number,
         @Query('item') item: number,
         @Query('difficulty') difficulty: RaidDifficulty,
@@ -30,8 +34,16 @@ export class WishlistsController {
     {
         if (!character)
             throw new BadRequestException('No character');
+
+        if (!teamId)
+            throw new BadRequestException('No team');
+        
+        const team = await this.teamsService.findById(teamId);
+        if (!team)
+            throw new NotFoundException('Team not found');
         
         return await this.wishlistsService.findByParams({
+            team: team,
             character: character,
             item: item,
             difficulty: difficulty,
@@ -42,22 +54,31 @@ export class WishlistsController {
 
     @UseGuards(JwtAuthenticationGuard)
     @ApiOperation({ summary: 'Toggle need of an item' })
+    @ApiQuery({ name: 'team', type: 'number' })
     @Post()
-    async toggle(@Body() wishDto: WishDto): Promise<boolean>
+    async toggle(@Query('team') teamId: number, @Body() wishDto: WishDto): Promise<boolean>
     {
+        if (!teamId)
+            throw new BadRequestException('No team');
+        
+        const team = await this.teamsService.findById(teamId);
+        if (!team)
+            throw new NotFoundException('Team not found');
+
         const needRes = await this.wishlistsService.findByParams({
             character: wishDto.character,
             item: wishDto.item,
-            difficulty: wishDto.difficulty
+            difficulty: wishDto.difficulty,
+            team: team
         })
 
         if (needRes) {
-            const need = needRes[0]; // They shouldn't be another one
+            const need = needRes[0]; // There shouldn't be another one
             return this.wishlistsService.delete(need.id);
         }
         else
         {
-            return (this.wishlistsService.create(wishDto) != undefined)
+            return (this.wishlistsService.save(wishDto) != undefined)
         }
     }
 }
